SearchPlus: Add UI and app communication
Add the UI for the application. This is most likely final unless
other changes are needed.

The app uses UserDefaults and a generic SearchEngine scheme to provide
both default search engines and custom search engines to the
extension with minimal overhead.

App groups are used to share UserDefaults between the app and its
extension.

Signed-off-by: kingbri <bdashore3@gmail.com>