WebExtension: Fix Yahoo issues
Yahoo does not follow the standard query paramters and uses 'p'
instead of 'q' inside the URL to provide a query.

Rather than running an operation and falling through, send
the entire user-set object to the JS script and if the system
engine's name is Yahoo, look for the 'p' URL parameter instead of 'q'.

Signed-off-by: kingbri <bdashore3@gmail.com>